+++ client/src/components/error/ErrorHooks.js
@@ -1,0 +1,9 @@
+import { useContext } from 'react';
+import { ErrorContext } from './ErrorProvider.js';
+
+function useGlobalError() {
+    const { error, setError, clearError } = useContext(ErrorContext);
+    return { error, setError, clearError };
+};
+
+export default useGlobalError;