+++ client/src/components/util/PaginationOptions.js
@@ -1,0 +1,10 @@
+const itemsPerPageOptions = [
+    { key: '10', text: '10', value: 10 },
+    { key: '25', text: '25', value: 25 },
+    { key: '50', text: '50', value: 50 },
+    { key: '100', text: '100', value: 100}
+];
+
+module.exports = {
+    itemsPerPageOptions
+}